requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
pydantic-settings
apscheduler
yfinance
pandas-ta
openai
httpx
python-telegram-bot[job-queue]
fastmcp
pandas_market_calendars
pytest
pytest-asyncio
